Add CardHistory and proper undoing using states
CardHistory is an object that constantly has copies of the model saved. Currently uncapped, though will be user settable, in the range of 3-10 ish. All the logic changes are in MainActivity, the rest of the commit is reflecting the change to using the getCard function across everything that accesses the model